Make code cleaner and more simple
